Move to `AuraNode`
